Simplify and deduplicate across branch changes
- Extract _X86_VARIANTS constant and move processor ambiguity checks
  to exceptions.py (shared between supervisor and worker)
- Short-circuit flag evaluation in _passes_type_filter
- Fix total_scanned to reflect actual items examined, not total available
- Make _build_instructions a @staticmethod
- Fix mid-word string break in instructions text
